Make routing key behaviour more clear in example
If I understand it correctly the routing-key in the policy is ignored by this exchange? Using the routing key "1234" confused me for a bit - it wasn't clear that it wasn't doing anything special.

I think it would make sense to call it "ignored", the fact it is ignored should be somewhat intuitive and if it isn't you use the word "ignored" in the documentation (making it easier to make the connection).